requirements.txt
anthropic==0.39.0
google-generativeai==0.8.3
numpy==2.1.3
openai==1.54.4
pandas==2.2.3
python==3.13